Handle admin credentials though parameterization
In addition to credentials-as-fact, allow them to be supplied by parameter.
Also, add a parameter to control whether the node should create the admin user,
so that in the case whether credentials are not available from any source there
can be a process that takes place on the stanchion node to create it (where
create_admin should be true) and a failure elsewhere until the resulting
credentials can be transmitted to an appropriate place.  Once established, the
credentials should be permanent unless all of the riak data is lost, which the
service is supposed to be hardened against.